# feat: add Scaled() method for scaled fields in mesgdef and reduce sec/op on ToMesg