--- a/profile/mesgdef/accelerometer_data_gen.go
+++ b/profile/mesgdef/accelerometer_data_gen.go
@@ -72,10 +72,10 @@ func NewAccelerometerData(mesg *proto.Message) *AccelerometerData {
 
 // ToMesg converts AccelerometerData into proto.Message.
 func (m *AccelerometerData) ToMesg(fac Factory) proto.Message {
-	fieldsPtr := fieldsPool.Get().(*[256]proto.Field)
-	defer fieldsPool.Put(fieldsPtr)
+	fieldsArray := fieldsPool.Get().(*[256]proto.Field)
+	defer fieldsPool.Put(fieldsArray)
 
-	fields := (*fieldsPtr)[:0] // Create slice from array with zero len.
+	fields := (*fieldsArray)[:0] // Create slice from array with zero len.
 	mesg := fac.CreateMesgOnly(typedef.MesgNumAccelerometerData)
 
 	if datetime.ToUint32(m.Timestamp) != basetype.Uint32Invalid {