--- a/profile/mesgdef/ant_channel_id_gen.go
+++ b/profile/mesgdef/ant_channel_id_gen.go
@@ -56,10 +56,10 @@ func NewAntChannelId(mesg *proto.Message) *AntChannelId {
 
 // ToMesg converts AntChannelId into proto.Message.
 func (m *AntChannelId) ToMesg(fac Factory) proto.Message {
-	fieldsPtr := fieldsPool.Get().(*[256]proto.Field)
-	defer fieldsPool.Put(fieldsPtr)
+	fieldsArray := fieldsPool.Get().(*[256]proto.Field)
+	defer fieldsPool.Put(fieldsArray)
 
-	fields := (*fieldsPtr)[:0] // Create slice from array with zero len.
+	fields := (*fieldsArray)[:0] // Create slice from array with zero len.
 	mesg := fac.CreateMesgOnly(typedef.MesgNumAntChannelId)
 
 	if m.ChannelNumber != basetype.Uint8Invalid {
@@ -67,24 +67,24 @@ func (m *AntChannelId) ToMesg(fac Factory) proto.Message {
 		field.Value = m.ChannelNumber
 		fields = append(fields, field)
 	}
-	if typeconv.ToUint8z[uint8](m.DeviceType) != basetype.Uint8zInvalid {
+	if uint8(m.DeviceType) != basetype.Uint8zInvalid {
 		field := fac.CreateField(mesg.Num, 1)
-		field.Value = typeconv.ToUint8z[uint8](m.DeviceType)
+		field.Value = uint8(m.DeviceType)
 		fields = append(fields, field)
 	}
-	if typeconv.ToUint16z[uint16](m.DeviceNumber) != basetype.Uint16zInvalid {
+	if uint16(m.DeviceNumber) != basetype.Uint16zInvalid {
 		field := fac.CreateField(mesg.Num, 2)
-		field.Value = typeconv.ToUint16z[uint16](m.DeviceNumber)
+		field.Value = uint16(m.DeviceNumber)
 		fields = append(fields, field)
 	}
-	if typeconv.ToUint8z[uint8](m.TransmissionType) != basetype.Uint8zInvalid {
+	if uint8(m.TransmissionType) != basetype.Uint8zInvalid {
 		field := fac.CreateField(mesg.Num, 3)
-		field.Value = typeconv.ToUint8z[uint8](m.TransmissionType)
+		field.Value = uint8(m.TransmissionType)
 		fields = append(fields, field)
 	}
-	if typeconv.ToUint8[uint8](m.DeviceIndex) != basetype.Uint8Invalid {
+	if uint8(m.DeviceIndex) != basetype.Uint8Invalid {
 		field := fac.CreateField(mesg.Num, 4)
-		field.Value = typeconv.ToUint8[uint8](m.DeviceIndex)
+		field.Value = uint8(m.DeviceIndex)
 		fields = append(fields, field)
 	}