--- a/profile/mesgdef/ant_tx_gen.go
+++ b/profile/mesgdef/ant_tx_gen.go
@@ -9,6 +9,7 @@ package mesgdef
 
 import (
 	"github.com/muktihari/fit/kit/datetime"
+	"github.com/muktihari/fit/kit/scaleoffset"
 	"github.com/muktihari/fit/kit/typeconv"
 	"github.com/muktihari/fit/profile/basetype"
 	"github.com/muktihari/fit/profile/typedef"
@@ -18,12 +19,12 @@ import (
 
 // AntTx is a AntTx message.
 type AntTx struct {
-	Timestamp           time.Time // Units: s;
-	FractionalTimestamp uint16    // Scale: 32768; Units: s;
+	Timestamp           time.Time // Units: s
+	FractionalTimestamp uint16    // Scale: 32768; Units: s
 	MesgId              byte
-	MesgData            []byte // Array: [N];
+	MesgData            []byte // Array: [N]
 	ChannelNumber       uint8
-	Data                []byte // Array: [N];
+	Data                []byte // Array: [N]
 
 	// Developer Fields are dynamic, can't be mapped as struct's fields.
 	// [Added since protocol version 2.0]
@@ -68,10 +69,10 @@ func NewAntTx(mesg *proto.Message) *AntTx {
 
 // ToMesg converts AntTx into proto.Message.
 func (m *AntTx) ToMesg(fac Factory) proto.Message {
-	fieldsPtr := fieldsPool.Get().(*[256]proto.Field)
-	defer fieldsPool.Put(fieldsPtr)
+	fieldsArray := fieldsPool.Get().(*[256]proto.Field)
+	defer fieldsPool.Put(fieldsArray)
 
-	fields := (*fieldsPtr)[:0] // Create slice from array with zero len.
+	fields := (*fieldsArray)[:0] // Create slice from array with zero len.
 	mesg := fac.CreateMesgOnly(typedef.MesgNumAntTx)
 
 	if datetime.ToUint32(m.Timestamp) != basetype.Uint32Invalid {
@@ -115,9 +116,19 @@ func (m *AntTx) ToMesg(fac Factory) proto.Message {
 	return mesg
 }
 
+// FractionalTimestampScaled return FractionalTimestamp in its scaled value [Scale: 32768; Units: s].
+//
+// If FractionalTimestamp value is invalid, float64 invalid value will be returned.
+func (m *AntTx) FractionalTimestampScaled() float64 {
+	if m.FractionalTimestamp == basetype.Uint16Invalid {
+		return basetype.Float64InvalidInFloatForm()
+	}
+	return scaleoffset.Apply(m.FractionalTimestamp, 32768, 0)
+}
+
 // SetTimestamp sets AntTx value.
 //
-// Units: s;
+// Units: s
 func (m *AntTx) SetTimestamp(v time.Time) *AntTx {
 	m.Timestamp = v
 	return m
@@ -125,7 +136,7 @@ func (m *AntTx) SetTimestamp(v time.Time) *AntTx {
 
 // SetFractionalTimestamp sets AntTx value.
 //
-// Scale: 32768; Units: s;
+// Scale: 32768; Units: s
 func (m *AntTx) SetFractionalTimestamp(v uint16) *AntTx {
 	m.FractionalTimestamp = v
 	return m
@@ -139,7 +150,7 @@ func (m *AntTx) SetMesgId(v byte) *AntTx {
 
 // SetMesgData sets AntTx value.
 //
-// Array: [N];
+// Array: [N]
 func (m *AntTx) SetMesgData(v []byte) *AntTx {
 	m.MesgData = v
 	return m
@@ -153,7 +164,7 @@ func (m *AntTx) SetChannelNumber(v uint8) *AntTx {
 
 // SetData sets AntTx value.
 //
-// Array: [N];
+// Array: [N]
 func (m *AntTx) SetData(v []byte) *AntTx {
 	m.Data = v
 	return m