--- a/profile/mesgdef/beat_intervals_gen.go
+++ b/profile/mesgdef/beat_intervals_gen.go
@@ -54,10 +54,10 @@ func NewBeatIntervals(mesg *proto.Message) *BeatIntervals {
 
 // ToMesg converts BeatIntervals into proto.Message.
 func (m *BeatIntervals) ToMesg(fac Factory) proto.Message {
-	fieldsPtr := fieldsPool.Get().(*[256]proto.Field)
-	defer fieldsPool.Put(fieldsPtr)
+	fieldsArray := fieldsPool.Get().(*[256]proto.Field)
+	defer fieldsPool.Put(fieldsArray)
 
-	fields := (*fieldsPtr)[:0] // Create slice from array with zero len.
+	fields := (*fieldsArray)[:0] // Create slice from array with zero len.
 	mesg := fac.CreateMesgOnly(typedef.MesgNumBeatIntervals)
 
 	if datetime.ToUint32(m.Timestamp) != basetype.Uint32Invalid {