- Reduce float32 and float64 invalid value creation, instead define as read-only global variable.
- Add method Scaled() for fields that can be scaled in mesgdef
- Reduce type conversion overhead by directly casting the type without using typeconv in ToMesg() method.
   benchstat:
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/profile/mesgdef
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
               │ recordmesg_old.txt │          recordmesg_new.txt          │
               │       sec/op       │    sec/op     vs base                │
RecordToMesg-4          195.9n ± 7%   144.9n ± 32%  -26.01% (p=0.001 n=10)

               │ recordmesg_old.txt │       recordmesg_new.txt       │
               │        B/op        │    B/op     vs base            │
RecordToMesg-4           8.000 ± 0%   8.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

               │ recordmesg_old.txt │       recordmesg_new.txt       │
               │     allocs/op      │ allocs/op   vs base            │
RecordToMesg-4           2.000 ± 0%   2.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

   ```